Converted client encryption response packet
